+++ src/Raycynix.Extensions.Database.Abstractions/IDatabaseProviderRegistration.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
 using Microsoft.EntityFrameworkCore;
-using Raycynix.Extensions.Database.Configurations;
+using Raycynix.Extensions.Database.Abstractions.Configurations;
 
 namespace Raycynix.Extensions.Database.Abstractions;