+++ src/Raycynix.Extensions.Database.MsSql/Internal/MsSqlServerDatabaseProviderRegistration.cs
@@ -3,7 +3,7 @@ using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 using Raycynix.Extensions.Configuration.Abstractions.Interfaces;
 using Raycynix.Extensions.Database.Abstractions;
-using Raycynix.Extensions.Database.Configurations;
+using Raycynix.Extensions.Database.Abstractions.Configurations;
 using Raycynix.Extensions.Database.MsSql.Configurations;
 
 namespace Raycynix.Extensions.Database.MsSql.Internal;