+++ src/Raycynix.Extensions.Database.PostgreSql/Internal/PostgreSqlDatabaseProviderRegistration.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using Npgsql;
 using Raycynix.Extensions.Configuration.Abstractions.Interfaces;
 using Raycynix.Extensions.Database.Abstractions;
-using Raycynix.Extensions.Database.Configurations;
+using Raycynix.Extensions.Database.Abstractions.Configurations;
 using Raycynix.Extensions.Database.PostgreSql.Configurations;
 
 namespace Raycynix.Extensions.Database.PostgreSql.Internal;