+++ src/Raycynix.Extensions.Database.Sqlite/Internal/SqliteDatabaseProviderRegistration.cs
@@ -3,7 +3,7 @@ using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
 using Raycynix.Extensions.Configuration.Abstractions.Interfaces;
 using Raycynix.Extensions.Database.Abstractions;
-using Raycynix.Extensions.Database.Configurations;
+using Raycynix.Extensions.Database.Abstractions.Configurations;
 using Raycynix.Extensions.Database.Sqlite.Configurations;
 
 namespace Raycynix.Extensions.Database.Sqlite.Internal;